pyproject.toml
[tool.poetry]
name = "bank-skypro"
version = "0.1.0"
description = ""
authors = ["aleksandr <aleksospishev@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pip = "^24.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
pytest = "^8.3.2"


[tool.poetry.group.lint.dependencies]
mypy = "^1.11.1"
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'


[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
    | venv
  )/
  | foo.py

)
'''

[tool.isort]
line_length = 119